Created abstract report class so we can have different types of compliance report without repetition.  Created time utility function. Cleanup.
