+++ src/main/java/edu/harvard/h2ms/service/report/AbstractReportWorkerComplianceWarning.java
@@ -1,4 +1,4 @@
-package edu.harvard.h2ms.service;
+package edu.harvard.h2ms.service.report;
 
 import edu.harvard.h2ms.domain.core.Event;
 import edu.harvard.h2ms.domain.core.Question;
@@ -6,9 +6,12 @@ import edu.harvard.h2ms.domain.core.User;
 import edu.harvard.h2ms.exception.InvalidAnswerTypeException;
 import edu.harvard.h2ms.repository.QuestionRepository;
 import edu.harvard.h2ms.repository.UserRepository;
+import edu.harvard.h2ms.service.EventService;
 import edu.harvard.h2ms.service.utils.H2msRestUtils;
 import edu.harvard.h2ms.service.utils.ReportUtils;
+import edu.harvard.h2ms.service.utils.ReportUtils.NotificationFrequency;
 import java.util.ArrayList;
+import java.util.Date;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
@@ -17,12 +20,10 @@ import org.hibernate.Hibernate;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.stereotype.Component;
 
-@Component
-public class ReportWorkerComplianceWarning implements ReportWorker {
+public abstract class AbstractReportWorkerComplianceWarning implements ReportWorker {
 
-  final Logger log = LoggerFactory.getLogger(ReportWorkerComplianceWarning.class);
+  final Logger log = LoggerFactory.getLogger(AbstractReportWorkerComplianceWarning.class);
 
   @Autowired private EventService eventService;
 
@@ -30,11 +31,14 @@ public class ReportWorkerComplianceWarning implements ReportWorker {
 
   @Autowired private UserRepository userRepository;
 
+  long REPORTINGINTERVAL = NotificationFrequency.WEEKLY.seconds;
+
   @Override
   public String getType() {
-    return "complianceWarning";
+    return "abstractComplianceWarning";
   }
 
+  @Override
   public String createReport() {
     List<Event> events = new ArrayList<>();
 
@@ -47,13 +51,24 @@ public class ReportWorkerComplianceWarning implements ReportWorker {
       try {
         events = eventService.findEventsForCompliance(question);
 
+        List<Event> timedEvents = new ArrayList<>();
+
+        for (Event event : events) {
+          Date eventDate = event.getTimestamp();
+          long unixtime = eventDate.getTime();
+          long reportTime = ReportUtils.getUnixTime();
+          if ((reportTime - unixtime) < (reportTime - REPORTINGINTERVAL)) {
+            timedEvents.add(event);
+          }
+        }
+
         // get compliance rates for all users:
         for (User user : userRepository.findAll()) {
           complianceResult.put(
               user,
               H2msRestUtils.calculateCompliance(
                   question,
-                  events
+                  timedEvents
                       .stream()
                       .filter(event -> event.getSubject().equals(user))
                       .collect(Collectors.toSet())));
@@ -84,5 +99,4 @@ public class ReportWorkerComplianceWarning implements ReportWorker {
 
     return ReportUtils.writeCsvString(data);
   }
-
 }
