+++ src/main/java/edu/harvard/h2ms/service/report/ReportService.java
@@ -1,4 +1,4 @@
-package edu.harvard.h2ms.service;
+package edu.harvard.h2ms.service.report;
 
 public interface ReportService {
