+++ src/main/java/edu/harvard/h2ms/service/report/ReportWorker.java
@@ -1,9 +1,8 @@
-package edu.harvard.h2ms.service;
+package edu.harvard.h2ms.service.report;
 
 public interface ReportWorker {
 
   public String getType();
 
   public String createReport();
-
 }
