+++ src/main/java/edu/harvard/h2ms/service/report/ReportWorkerComplianceWarningMonthly.java
@@ -1,0 +1,3 @@
+package edu.harvard.h2ms.service.report;
+
+public class ReportWorkerComplianceWarningMonthly {}
