+++ src/main/java/edu/harvard/h2ms/web/controller/EventController.java
@@ -6,7 +6,7 @@ import edu.harvard.h2ms.exception.InvalidAnswerTypeException;
 import edu.harvard.h2ms.exception.InvalidTimeframeException;
 import edu.harvard.h2ms.repository.QuestionRepository;
 import edu.harvard.h2ms.service.EventService;
-import edu.harvard.h2ms.service.ReportService;
+import edu.harvard.h2ms.service.report.ReportService;
 import java.util.List;
 import java.util.Map;
 import org.slf4j.Logger;
